finish refactoring
