+++ client/src/components/content/ModelerSidebar/ModelerSidebar.jsx
@@ -3,6 +3,8 @@ import React from 'react';
 import { Layout, Button, Space } from 'antd';
 import styles from './ModelerSidebar.module.css';
 import PropertiesPanel from '../PropertiesPanel/PropertiesPanel';
+import getParsedRobotFile from "../../../api/ssot";
+import downloadString from '../../../utils/downloadString';
 
 const { Sider } = Layout;
 
@@ -11,15 +13,29 @@ const { Sider } = Layout;
  * @category Client
  * @component
  */
-const ModelerSidebar = ({ modeler, getRobotFile }) => (
-  <Sider className={styles.sider}>
-    <Space direction='vertical' size='large' style={{ width: '100%' }}>
-      {modeler && <PropertiesPanel modeler={modeler} />}
-      <Button type='primary' className={styles.button} onClick={getRobotFile}>
-        Get Robot file
+const ModelerSidebar = ({ modeler, getRobotFile }) => {
+  /**
+ * @description Will parse a given xml file into a .robot file and download it
+ * @param {string} xml String that sets the xml to be parsed
+ */
+  const downloadRobotFile = () => {
+    getParsedRobotFile()
+      .then((response) => response.text())
+      .then((robotCode) => {
+        downloadString(robotCode, 'text/robot', 'testRobot.robot');
+      });
+  };
+  return (
+    <Sider className={styles.sider}>
+      <Space direction='vertical' size='large' style={{ width: '100%' }}>
+        {modeler && <PropertiesPanel modeler={modeler} />}
+        <Button type='primary' className={styles.button} onClick={downloadRobotFile}>
+          Get Robot file
       </Button>
-    </Space>
-  </Sider>
-);
+      </Space>
+    </Sider>
+
+  )
+}
 
 export default ModelerSidebar;
